refactor(chunker): split chunker.rs into module skeleton with legacy shim
