+++ src/transform/chunker/engine.rs
@@ -1,0 +1,1 @@
+//! Placeholder filled by Task 5.
