+++ src/transform/chunker/fingerprint.rs
@@ -1,0 +1,5 @@
+//! Placeholder filled by Task 3.
+
+/// Stub: real definition comes in Task 3.
+#[derive(Debug, Clone, Default)]
+pub struct StrategyFingerprint;
