+++ src/transform/chunker/recursive.rs
@@ -1,0 +1,5 @@
+//! Placeholder filled by Task 6.
+
+/// Stub: real definition comes in Task 6.
+#[derive(Debug, Clone, Default)]
+pub struct RecursiveChunker;
