+++ src/transform/chunker/size.rs
@@ -1,0 +1,19 @@
+//! Placeholder filled by Task 4.
+
+/// Stub: real definition comes in Task 4.
+pub trait TokenCounter: Send + Sync {}
+
+/// Stub: real definition comes in Task 4.
+#[derive(Debug, Clone, Default)]
+pub struct CharCounter;
+
+/// Stub: real definition comes in Task 4.
+#[derive(Debug, Clone, Default)]
+pub struct TiktokenCounter;
+
+/// Stub: real definition comes in Task 4.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SizeMetric {
+    Chars,
+    Tokens,
+}
